// src/shared/api/mold-passport/model/moldingAreaLookupResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Molding Shop Service
 * OpenAPI spec version: 0.1.0
 */
import type { MoldingAreaLookupResponsePressureUnits } from './moldingAreaLookupResponsePressureUnits'

export interface MoldingAreaLookupResponse {
  id: number
  name: string
  pressureUnits?: MoldingAreaLookupResponsePressureUnits
}