// src/shared/api/mold-passport/model/moldingAreaLookupsListResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Molding Shop Service
 * OpenAPI spec version: 0.1.0
 */
import type { MoldingAreaLookupResponse } from './moldingAreaLookupResponse'

export interface MoldingAreaLookupsListResponse {
  data: MoldingAreaLookupResponse[]
  page: number
  totalPages: number
  totalItems: number
}